Added tilemap to shared class
