+++ Client/Client/Rendering/WorldMap.cs
@@ -1,74 +1,24 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Text.Json;
-using System.Text.Json.Serialization;
+using SharedLibrary;
 
 namespace Client.Rendering
 {
-    public class Tilemap
+    public class WorldMap : Tilemap
     {
         private Texture2D tilesetTexture;
         private Rectangle[] tileSourceRects;
-        private JsonSerializerOptions options;
-
-        public string Name { get; set; }
-        public string TexturePath { get; set; }
-        public int TileSize { get; set; }
-        public int TilesetWidth { get; set; }
-        public int TilesetHeight { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
-        public List<TileProperty> TilesetData { get; set; }
-        public int[][] Tiles { get; set; }
-
-        public Tilemap()
-        { 
-            options = new JsonSerializerOptions
-            {
-                PropertyNameCaseInsensitive = true,
-                Converters =
-                {
-                    new JsonStringEnumConverter()
-                }
-            };
-        }
 
         /// <summary>
         /// Loads the tilemap from a JSON file and sets the current map data.
         /// </summary>
         /// <param name="filePath">The path to the JSON file containing the map data.</param>
         /// <returns>True if the map was loaded successfully; otherwise, false.</returns>
-        public bool LoadMap(string filePath, ContentManager contentManager)
+        public bool InitMap(string filePath, ContentManager contentManager)
         {
-            try
+            if (LoadMap(filePath))
             {
-                if (!File.Exists(filePath))
-                {
-                    Console.WriteLine($"[LoadMap] File not found: {filePath}");
-                    return false;
-                }
-
-                string json = File.ReadAllText(filePath);
-                Tilemap loadedMap = JsonSerializer.Deserialize<Tilemap>(json, options);
-
-                if (loadedMap == null)
-                {
-                    Console.WriteLine($"[LoadMap] Failed to deserialize JSON: {filePath}");
-                    return false;
-                }
-
-                Name = loadedMap.Name;
-                Width = loadedMap.Width;
-                Height = loadedMap.Height;
-                Tiles = loadedMap.Tiles;
-                TexturePath = loadedMap.TexturePath;
-                TileSize = loadedMap.TileSize;
-                TilesetWidth = loadedMap.TilesetWidth;
-                TilesetHeight = loadedMap.TilesetHeight;
                 tilesetTexture = contentManager.Load<Texture2D>(TexturePath);
 
                 int tilesPerRow = TilesetWidth / TileSize;
@@ -86,9 +36,8 @@ namespace Client.Rendering
 
                 return true;
             }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"[LoadMap] Error loading map from {filePath}: {ex.Message}");
+            else
+            { 
                 return false;
             }
         }
